+++ .env.example
@@ -1,2 +1,2 @@
 TOKEN=
-DEV=true
+NODE_ENV="development"
