--- config.ts
@@ -1,24 +1,0 @@
-export default {
-  clientURL: "https://chat.fribyte.no",
-  happenings:  [
-    {
-      title: "Dugnad",
-      daysToMessage: [1, 3], // Monday and Wednesday
-      timeOfDayToMessage: 17, // 5pm
-      messages: [
-        "Dugnad i morgen friBytere? :60fpsparrot: \n// Dugnad tomorrow friByters?",
-        "Dugnad i morgen mennesker og roboter? \n// Dugnad tomorrow humans and robots?",
-      ],
-    },
-    //TODO - see issue #1
-    // {
-    //   title: "Social",
-    //   daysToMessage: [0, 2], // Sunday and Tuesday
-    //   timeOfDayToMessage: 17, // 5pm
-    //   messages: [
-    //     "Dugnad i morgen friBytere? :60fpsparrot:",
-    //     "Dugnad i morgen mennesker og roboter?",
-    //   ],
-    // }
-  ]
-};
