+++ package.json
@@ -3,7 +3,7 @@
   "module": "index.ts",
   "type": "module",
   "scripts": {
-    "build": "bun run serve index.ts"
+    "start": "bun run src/index.ts"
   },
   "devDependencies": {
     "bun-types": "latest",
